% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{country}
\alias{country}
\title{Penn World Tables 9.0 Country List}
\format{A tibble with 182 rows and 2 variables:
\describe{
  \item{countrycode}{countrycode}
  \item{country}{country}
}}
\usage{
country
}
\description{
Penn World Tables 9.0 Country List
}
\keyword{datasets}
