% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_lag_ops.R
\name{D}
\alias{D}
\title{Difference of a Time Series}
\usage{
D(x, lag = 1, diff = 1)
}
\arguments{
\item{x}{`xts` time series variable}

\item{lag}{numeric or character value indicating the number of periods to
lag x. Defaults to one lag. Character inputs indicate the type of time
period to lag. \code{lag} accepts 'y' years, 'q' quarters, 'm' months,
'w' weeks, and 'd' days. Number and character values can be combined.}

\item{diff}{number of times the \code{x} is differenced}
}
\value{
differenced 'xts' object.
}
\description{
\code{D} calculates the difference between an \code{xts} time series and
its lag.
}
\examples{
Y <- pwt(country = 'USA', series = 'gdp')
d_Y <- D(Y)           # GDP differenced by 1 period (quarter)
y_Y <- D(Y, 'y')      # GDP differenced by 1 year
D_Y <- D(Y, diff = 2) # GDP differenced twice
#Table(Y, d_y, y_Y, D_Y)
}

