% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fred}
\alias{fred}
\title{Download Federal Reserve Data}
\usage{
fred(..., names = NULL, rm = TRUE, log = NULL, as_xts = FALSE)
}
\arguments{
\item{...}{vector of FRED codes as character strings (use quotations).}

\item{names}{vector of characters to replace column names in final dataframe. The length must be equal to the number of codes.}

\item{rm}{logical value expressing whether or not to remove time periods that contain an NA value. Defaults to TRUE.}

\item{log}{character vector describing which variables to transform using the natural logarithm.}

\item{as_xts}{logical value describing whether the output should be 'xts' type.}
}
\value{
dataframe
}
\description{
\code{fred} downloads data from the Federal Reserve Economic Database
(FRED).
}
\examples{
y <- fred('GDP')
macro <- fred('GDP', 'CPIAUCSL', 'UNRATE', names = c('y', 'pi', 'u'))

}
