% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_functions.R
\name{fredSearch}
\alias{fredSearch}
\title{Search Federal Reserve Data}
\usage{
fredSearch(x, ret = "obs", n = 30)
}
\arguments{
\item{x}{character value to search}

\item{ret}{either 'obs' or 'id' to return either the data observations
or the FRED id code}

\item{n}{the maximum number of results to show per page}
}
\value{
'xts' object or a FRED code
}
\description{
\code{fredSearch} searches and downloads data from the Federal Reserve
Economic Database.
}
\examples{
urate <- fredSearch('unemp rate')
medPrice <- fredSearch('medical care price')
}

