% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recon_functions.R
\name{freq}
\alias{freq}
\title{Convert to Specific Frequency}
\usage{
freq(x, freq = "q", OHLC = FALSE)
}
\arguments{
\item{x}{'xts' object}

\item{freq}{frequency which to convert. Accepts 'y', 'q', 'm', 'w',
or 'd' for dates. Also accepts 'h', '30m', '15m', '10m', '5m', '3m',
and 'mn' for times.}

\item{OHLC}{logical describing whether to keep the Open, High, Low, and
Close of data after conversion. Defaults to FALSE.}
}
\value{
'xts' object
}
\description{
\code{freq} converts an 'xts' time series object to a different time
frequency.
}
\examples{
urate <- fred('UNRATE', as_xts=TRUE) # monthly Civilian Unemployment Rate data
freq(urate, 'y')        # yearly Unemployment Rate
}
