% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recon_functions.R
\name{G}
\alias{G}
\title{Growth Rate of a Time Series}
\usage{
G(x, lag = 1)
}
\arguments{
\item{x}{`xts` time series variable}

\item{lag}{numeric or character value indicating the number of periods
between which the time series grew}
}
\value{
'xts' object of growth rates.
}
\description{
\code{G} calculates the growth rate of an 'xts' time series.
}
\examples{
Y <- pwt(country = 'USA', series = 'gdp')
g_Y <- G(Y)      # 1 period growth rate of GDP
y_Y <- G(Y, 'y') # yearly growth rate of GDP
m_Y <- G(Y, '6m') # 6 month growth rate of GDP
#Table(Y, g_y, y_Y, m_Y)
}
