% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hw1}
\alias{hw1}
\title{Homework 1 Data Set}
\format{A tibble with 282 rows and 3 variables:
\describe{
  \item{Date}{Date}
  \item{Cons}{Real personal consumption expenditures per capita}
  \item{Price}{Consumer Price Index for All Urban Consumers: All Items}
}}
\usage{
hw1
}
\description{
Homework 1 Data Set
}
\keyword{datasets}
