% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hw2}
\alias{hw2}
\title{Homework 2 Data Set}
\format{A tibble with 295 rows and 9 variables:
\describe{
  \item{Date}{Date}
  \item{Foo}{Consumer Price Index for All Urban Consumers: Food and Beverages}
  \item{Hou}{Consumer Price Index for All Urban Consumers: Housing}
  \item{App}{Consumer Price Index for All Urban Consumers: Apparel}
  \item{Tra}{Consumer Price Index for All Urban Consumers: Transportation}
  \item{Med}{Consumer Price Index for All Urban Consumers: Medical Care}
  \item{Rec}{Consumer Price Index for All Urban Consumers: Recreation}
  \item{Edu}{Consumer Price Index for All Urban Consumers: Education and Communication}
  \item{Oth}{Consumer Price Index for All Urban Consumers: Other Goods and Services}
}}
\usage{
hw2
}
\description{
Homework 2 Data Set
}
\keyword{datasets}
