% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hw5}
\alias{hw5}
\title{Homework 5 Data Set}
\format{A tibble with 234 rows and 4 variables:
\describe{
  \item{Date}{Date}
  \item{w}{Compensation of Employees, Received: Wage and Salary Disbursements}
  \item{u}{Civilian Unemployment Rate}
  \item{un}{Natural Rate of Unemployment (Long-Term)}
}}
\usage{
hw5
}
\description{
Homework 5 Data Set
}
\keyword{datasets}
