% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hw6}
\alias{hw6}
\title{Homework 6 Data Set}
\format{A tibble with 234 rows and 4 variables:
\describe{
  \item{Date}{Date}
  \item{P}{Consumer Price Index for All Urban Consumers: All Items}
  \item{M}{M2 Money Stock}
  \item{Y}{Real Gross Domestic Product}
}}
\usage{
hw6
}
\description{
Homework 6 Data Set
}
\keyword{datasets}
