% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hw7}
\alias{hw7}
\title{Homework 7 Data Set}
\format{A tibble with 257 rows and 6 variables:
\describe{
  \item{Date}{Date}
  \item{P}{Consumer Price Index for All Urban Consumers: All Items}
  \item{i}{10-Year Treasury Constant Maturity Rate}
  \item{Y}{Real Gross Domestic Product}
  \item{pY}{Real Potential Gross Domestic Product}
  \item{G}{Real Government Consumption Expenditures and Gross Investment}
}}
\usage{
hw7
}
\description{
Homework 7 Data Set
}
\keyword{datasets}
