% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hw8}
\alias{hw8}
\title{Homework 8 Data Set}
\format{A tibble with 230 rows and 4 variables:
\describe{
  \item{Date}{Date}
  \item{pi}{Inflation Rate from Personal Consumption Expenditures Chain-Type Price Index Less Food and Energy}
  \item{y}{GDP Gap}
  \item{i}{Federal Funds Rate}
}}
\usage{
hw8
}
\description{
Homework 8 Data Set
}
\keyword{datasets}
