% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{info}
\alias{info}
\title{Obtain Information About Time Series}
\usage{
info(x)
}
\arguments{
\item{x}{FRED variable}
}
\value{
list
}
\description{
\code{info} lists the information about a time series obtained via FRED.
}
\examples{
data(macro)
info(macro$GDP)

}
