% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recon_functions.R
\name{L}
\alias{L}
\title{Lag of a Time Series}
\usage{
L(x, lag = 1)
}
\arguments{
\item{x}{`xts` time series variable}

\item{lag}{numeric or character value indicating the number of periods to
lag x. Defaults to one lag. Character inputs indicate the type of time
period to lag. \code{lag} accepts 'y' years, 'q' quarters, 'm' months,
'w' weeks, and 'd' days. Number and character values can be combined.}
}
\value{
new 'xts' object shifted by the \code{lag} number of periods.
}
\description{
\code{L} shifts the values in an \code{xts} time series by the specified
number of lags. The function extends the \code{lag} function by allowing
users to specify a frequency with which to lag the variable.
}
\examples{
Y <- pwt(country = 'USA', series = 'gdp')
l_Y <- L(Y)       # GDP lagged by 1 period (quarter)
y_Y <- L(Y, 'y')  # GDP lagged by 1 year
m_Y <- L(Y, '6m') # GDP lagged by 6 months
#Table(Y, l_y, y_Y, m_Y)
}
