% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{labor}
\alias{labor}
\title{Labor Market Data}
\format{A tibble with 136 rows and 38 variables:
\describe{
  \item{Date}{Date}
  \item{rec}{NBER based Recession Indicators for the United States from the Period following the Peak through the Trough}
  \item{lmci}{Change in Labor Market Conditions Index (DISCONTINUED)}
  \item{u3}{Civilian Unemployment Rate}
  \item{u1}{Persons Unemployed 15 weeks or longer, as a percent of the civilian labor force}
  \item{u2}{Unemployment Rate: Job Losers}
  \item{u4}{Special Unemployment Rate: Unemployed and Discouraged Workers}
  \item{u5}{Special Unemployment Rate: Unemployed and Marginally Attached Workers}
  \item{u6}{Total unemployed, plus all marginally attached workers plus total employed part time for economic reasons}
  \item{unemp}{Unemployment Level}
  \item{emp}{Civilian Employment Level}
  \item{pr}{Civilian Labor Force Participation Rate}
  \item{pr.men}{Civilian Labor Force Participation Rate: Men}
  \item{pr.fem}{Civilian Labor Force Participation Rate: Women}
  \item{vac}{Job Openings: Total Private}
  \item{hires}{Hires: Total Private}
  \item{sep}{Total Separations: Total Private}
  \item{quits}{Quits: Total Private}
  \item{layoff}{Layoffs and Discharges: Total Private}
  \item{earn}{Average Hourly Earnings of All Employees: Total Private}
  \item{hours}{Average Weekly Hours of All Employees: Total Private}
  \item{hrns}{Average Hourly Earnings of Production and Nonsupervisory Employees: Total Private}
  \item{hrman}{Average Weekly Hours of Production and Nonsupervisory Employees: Manufacturing}
  \item{dur}{Average (Mean) Duration of Unemployment}
  \item{dur.med}{Median Duration of Unemployment}
  \item{em}{All Employees: Total Private Industries}
  \item{emcon}{All Employees: Construction}
  \item{emgov}{All Employees: Government}
  \item{emtrade}{All Employees: Retail Trade}
  \item{emwtrade}{All Employees: Wholesale Trade}
  \item{emfin}{All Employees: Financial Activities}
  \item{emmine}{All Employees: Mining and logging}
  \item{emehs}{All Employees: Education and Health Services}
  \item{empbs}{All Employees: Professional and Business Services}
  \item{eminf}{All Employees: Information Services}
  \item{emlh}{All Employees: Leisure and Hospitality}
  \item{emttu}{All Employees: Trade, Transportation and Utilities}
  \item{emman}{All Employees: Manufacturing}
}}
\usage{
labor
}
\description{
Labor Market Data
}
\keyword{datasets}
