% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{macro}
\alias{macro}
\title{General Macroeconomics Data}
\format{A tibble with 205 rows and 29 variables:
\describe{
  \item{Date}{Date}
  \item{rec}{NBER based Recession Indicators for the United States from the Period following the Peak through the Trough}
  \item{gdp}{Gross Domestic Product}
  \item{rgdp}{Real Gross Domestic Product}
  \item{pgdp}{Real Potential Gross Domestic Product}
  \item{deflator}{Gross Domestic Product: Implicit Price Deflator}
  \item{cpi}{Consumer Price Index for All Urban Consumers: All Items}
  \item{pce}{Personal Consumption Expenditures Excluding Food and Energy (Chain-Type Price Index)}
  \item{m2}{M2 Money Stock}
  \item{m2v}{Velocity of M2 Money Stock}
  \item{u}{Civilian Unemployment Rate}
  \item{un}{Natural Rate of Unemployment (Long-Term)}
  \item{emp}{Civilian Employment Level}
  \item{unemp}{Unemployment Level}
  \item{workpop}{Working Age Population: Aged 15-64: All Persons for the United States©}
  \item{ffr}{Effective Federal Funds Rate}
  \item{i10}{10-Year Treasury Constant Maturity Rate}
  \item{c}{Personal Consumption Expenditures}
  \item{i}{Gross Private Domestic Investment}
  \item{g}{Government Consumption Expenditures and Gross Investment}
  \item{ex}{Exports of Goods and Services}
  \item{im}{Imports of Goods and Services}
  \item{rc}{Real Personal Consumption Expenditures}
  \item{ri}{Real Gross Private Domestic Investment}
  \item{rg}{Real Government Consumption Expenditures and Gross Investment}
  \item{rex}{Real Exports of Goods and Services}
  \item{rim}{Real imports of goods and services}
  \item{debt}{Federal Debt: Total Public Debt}
  \item{pop}{Population (midperiod)}
}}
\usage{
macro
}
\description{
General Macroeconomics Data
}
\keyword{datasets}
