% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nipa}
\alias{nipa}
\title{National Income and Products Accounts}
\format{A tibble with 282 rows and 15 variables:
\describe{
  \item{Date}{Date}
  \item{rec}{NBER based Recession Indicators for the United States from the Period following the Peak through the Trough}
  \item{GDP}{Gross Domestic Product}
  \item{Deflator}{Gross Domestic Product: Implicit Price Deflator}
  \item{C}{Personal Consumption Expenditures}
  \item{I}{Gross Private Domestic Investment}
  \item{G}{Government Consumption Expenditures and Gross Investment}
  \item{EX}{Exports of Goods and Services}
  \item{IM}{Imports of Goods and Services}
  \item{dg}{Personal Consumption Expenditures: Durable Goods}
  \item{nd}{Personal Consumption Expenditures: Nondurable Goods}
  \item{sv}{Personal Consumption Expenditures: Services}
  \item{cbi}{Change in Private Inventories}
  \item{nres}{Private Nonresidential Fixed Investment}
  \item{res}{Private Residential Fixed Investment}
}}
\usage{
nipa
}
\description{
National Income and Products Accounts
}
\keyword{datasets}
