% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nk}
\alias{nk}
\title{New Keynesian Model Data}
\format{A tibble with 234 rows and 7 variables:
\describe{
  \item{Date}{Date}
  \item{Y}{Real Gross Domestic Product}
  \item{pY}{Real Potential Gross Domestic Product}
  \item{U}{Civilian Unemployment Rate}
  \item{Un}{Natural Rate of Unemployment (Long-Term)}
  \item{P}{Personal Consumption Expenditures Excluding Food and Energy (Chain-Type Price Index)}
  \item{FFR}{Effective Federal Funds Rate}
}}
\usage{
nk
}
\description{
New Keynesian Model Data
}
\keyword{datasets}
