% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recon_functions.R
\name{printDate}
\alias{printDate}
\title{Pretty Print Dates}
\usage{
printDate(x, freq, format = NULL)
}
\arguments{
\item{x}{character or Date value}

\item{freq}{frequency to print. Accepts 'y', 'q', 'm', or 'd'}

\item{format}{if \code{x} is a character, specifies the date format}
}
\value{
character value
}
\description{
\code{printDate} truncates and cleans dates values according to frequency.
}
\examples{
printDate('1947-01-01', freq='q')
printDate('1965-05-30', freq='m')
}
