% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwt.R
\name{pwt}
\alias{pwt}
\title{Extract PWT 9.0 Data}
\usage{
pwt(country, series)
}
\arguments{
\item{country}{3-digit ISO country code}

\item{series}{variable name to load}
}
\value{
'xts' object
}
\description{
\code{pwt} extracts and formats specific data from the Penn World Tables as
'xts' time series objects.
}
\examples{
Y <- pwt(country = 'USA', series = 'gdp')
tPlot(Y)
}

