% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwt.R
\docType{data}
\name{pwt9}
\alias{pwt9}
\title{Penn World Tables 9.0}
\format{A data frame with 8244 rows and 16 variables:
\describe{
  \item{code}{3-letter ISO country code}
  \item{year}{Year}
  \item{pop}{Population}
  \item{emp}{Labor Force}
  \item{hours}{Average Annual Hours of Labor Force}
  \item{hc}{Human Capital Index}
  \item{gdp}{Real GDP at Constant 2011 National Prices}
  \item{k}{Capital Stock at Constant 2011 National Prices}
  \item{tfp}{TFP at Constant National Prices (2011=1)}
  \item{labsh}{Share of Labour Compensation in GDP at Current Nat. Prices}
  \item{delta}{Average Depreciation Rate of the Capital Stock}
  \item{shC}{Share of Household Consumption at current PPPs}
  \item{shI}{Share of Gross Capital Formation at current PPPs}
  \item{shG}{Share of Government Consumption at current PPPs}
  \item{shX}{Share of Merchandise Exports at current PPPs}
  \item{shM}{Share of Merchandise Imports at current PPPs}
}}
\source{
\url{http://www.rug.nl/ggdc/productivity/pwt/}
}
\usage{
pwt9
}
\description{
A dataset containing data from the Penn World Tables version 9.0. It
includes data from 178 countries. This is not the complete PWT. Observations
are only included if they contain data on Real GDP, capital stock, and
employment. Furthermore, only 16 variables are included. More info about the
variables are included in the attributes.
}
\keyword{datasets}

