% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwt.R
\docType{data}
\name{pwtcountries}
\alias{pwtcountries}
\title{Penn World Tables 9.0 Countries}
\format{A data frame with 178 rows and 3 variables:
\describe{
  \item{Code}{3-letter ISO country code}
  \item{Country}{Country name}
  \item{Years}{Years available in the dataset}
}}
\source{
\url{http://www.rug.nl/ggdc/productivity/pwt/}
}
\usage{
pwtcountries
}
\description{
A dataset containing the countries in the PennWorld Table.
}
\keyword{datasets}

