% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwt.R
\name{pwtCountry}
\alias{pwtCountry}
\title{PWT 9.0 Country Names}
\usage{
pwtCountry(search = NULL)
}
\arguments{
\item{search}{character string to search for within country names. If NULL
lists all countries}
}
\value{
data.frame
}
\description{
\code{pwtCountry} lists available country codes for \code{pwt} function and
the years available
}
\examples{
pwtCountry()
pwtCountry('swed')
pwtCountry('ir')
}

