% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwt.R
\name{pwtSeries}
\alias{pwtSeries}
\title{PWT 9.0 Series Names}
\usage{
pwtSeries()
}
\value{
data.frame
}
\description{
\code{pwtSeries} lists available series for \code{pwt} function and info
}
\examples{
pwtSeries()
}

