% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg.R
\name{Reg}
\alias{Reg}
\title{Regression}
\usage{
Reg(y, ..., data = NULL, time = NULL, test = TRUE)
}
\arguments{
\item{data}{object that contains data in formula}

\item{form}{regression formula, e.g. \emph{y~x1+x2}}
}
\value{
'RegM' object containing regression information, including
  coefficients, se, tests, and others
}
\description{
\code{Reg} performs a variety of regressions within a single function (only
OLS is supported at the moment). Standard test are automatically performed.
Multiple types of standard error calculations are available.
}
\examples{
GDPgap <- fred('gap')
cycU <- fred('cyclical')
Reg(GDPgap ~ cycU) # Okun's Law
}

