% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{repChar}
\alias{repChar}
\title{Repeat Character in String}
\usage{
repChar(x, n, nl = FALSE)
}
\arguments{
\item{x}{characters to repeats}

\item{n}{number of times to repeat}

\item{nl}{logical determining whether a newline should be added}
}
\value{
character of length \code{nchar(x)*n}
}
\description{
\code{repChar} repeats a character for a single value.
}
\examples{
repChar('=', 10)
}

