% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rnipa}
\alias{rnipa}
\title{Real National Income and Products Accounts}
\format{A tibble with 282 rows and 8 variables:
\describe{
  \item{Date}{Date}
  \item{rec}{NBER based Recession Indicators for the United States from the Period following the Peak through the Trough}
  \item{Y}{Real Gross Domestic Product}
  \item{C}{Real Personal Consumption Expenditures}
  \item{I}{Real Gross Private Domestic Investment}
  \item{G}{Real Government Consumption Expenditures and Gross Investment}
  \item{X}{Real Exports of Goods and Services}
  \item{M}{Real imports of goods and services}
}}
\usage{
rnipa
}
\description{
Real National Income and Products Accounts
}
\keyword{datasets}
