% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sPlot.R
\name{splot}
\alias{splot}
\title{Scatter Plot Time Series}
\usage{
splot(x, y, by = NULL, data = NULL, title = NULL, xlab = NULL,
  ylab = NULL, time = NULL, ylim = NULL, xlim = NULL, type = "p",
  col = NULL, lwd = 1, lty = NULL, colAxes = NULL, pch = NULL,
  cex = 1, colBG = NULL, bg = colBG, colPlot = NULL, colGrid = NULL,
  theme = "std", reg = NULL, rlwd = lwd + 1, mar = c(4, 4, 2, 1),
  cex.lab = 1, cex.axis = 0.75, las = 0, leg = NULL)
}
\arguments{
\item{x}{'xts' object to be plotted on the x-axis}

\item{y}{'xts' object to be plotted on the y-axis}

\item{by}{object of factors to be plotted. Only 'cycle' supported for now,
which plots by the business cycle}

\item{data}{data object containing \code{x} and \code{x}}

\item{title}{character title of plot}

\item{xlab}{character label of x-axis}

\item{ylab}{character label of y-axis}

\item{time}{time period to restrict plot. Uses 'xts' convention, e.g.,
'1970-01/1972-06-03'.}

\item{ylim}{numeric vector of limits on the y-axis}

\item{xlim}{Date vector of limits on the x-axis}

\item{type}{'p' for points, 'l' for lines, 'b' for both}

\item{col}{character vector of color names to use for lines and points}

\item{lwd}{line width}

\item{lty}{line type. examples: 'solid', 'dotted', 'dashed'}

\item{colAxes}{color of plot axes}

\item{pch}{type of point to use}

\item{cex}{size of objects in plot}

\item{colBG}{background color}

\item{bg}{background color of points}

\item{colPlot}{color of plot background}

\item{colGrid}{color of grids}

\item{theme}{graphical theme of the plot. Only 'std' is supported.}

\item{reg}{logical or character. Should a regression line by printed? If
'by', then a regression for each category.}

\item{rlwd}{regression line width}

\item{mar}{inner margins outside of plot. c(bottom,left,top,right)}

\item{cex.lab}{size of labels}

\item{cex.axis}{size of axis}

\item{las}{tick label orientation on axis}

\item{leg}{vector of character names to use in the legend. 'auto' uses
object attributes}
}
\value{
plot of time-series
}
\description{
\code{splot} plots two economic time series against each other, taking into
consideration object attributes such as 'units' and 'title'. splot works
best with data collecting functions, such as \code{fred} in the the
\code{recon} package.
}
\section{Graphical Parameters}{

If \code{NULL}, then value is chosen either by \code{theme} or with object
attributes.
}
\examples{
GDPgap <- fred('gap')
cycU <- fred('cyclical')
splot(GDPgap, cycU, reg=T)
splot(GDPgap, cycU, by='cycle')
}

