% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recon_functions.R
\name{Table}
\alias{Table}
\title{Pretty Print Data}
\usage{
Table(x, n = 30, align = "auto", digits = 3, trunc = "long",
  rows = TRUE, sub = NULL, start = 1, sep = NULL)
}
\arguments{
\item{x}{object to be printed}

\item{n}{maximum number of rows to print. If 'ht', head and tail are printed}

\item{align}{vector of characters describing column alignment. Accepts
'l' for left, 'r' for right, 'c' for center, 'd' for aligning on
decimal place, 't' for left alignment and truncation, 'm' for truncating
to the shortest character. 'auto' aligns center for characters and on
the decimal place for numeric values.}

\item{digits}{number of digits to which numeric values are rounded}

\item{trunc}{character or numeric value. Character values of 'long' or
'col' are accepted. 'long' truncates the widest column if necessary,
'col' removes necessary columns. Numeric values specify which columns to
truncate.}

\item{rows}{logical. Are row numbers printed?}

\item{sub}{vector to be printed under column names. If 'type', the type of
each column is printed}

\item{start}{starting row number of the data to print}

\item{sep}{defaults to NULL. Row number after which to draw a line}
}
\value{
prints truncated table of data
}
\description{
\code{Table} formats and prints data. Rows and columns are truncated to
fit on screen and columns are aligned.
}
\examples{
GDP <- fred('GDP')
Table(GDP)

data(pwt9)
Table(pwt9, sub = 'type', trunc = 'col')
}
