% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tPlot.R
\name{tPlot}
\alias{tPlot}
\title{Plot Time Series}
\usage{
tPlot(..., right = NULL, data = NULL, time = NULL, xlim = NULL,
  ylim = NULL, yRlim = NULL, ticks = c(8, 5), theme = "std",
  col = NULL, colBG = NULL, colGrid = NULL, colAxes = NULL,
  colPlot = NULL, colBar = NULL, mar = c(3.5, 3.5, 2, 1),
  bars = "cycle", las = 0, cex = 1, cex.lab = 1, cex.axis = 0.75,
  lty = NULL, lwd = 1, xlab = "Date", ylab = NULL, yRlab = NULL,
  title = NULL, ltyAxis = "solid", lwdAxis = 1, lwdTicks = lwdAxis,
  leg = "auto")
}
\arguments{
\item{...}{'xts' objects to be plotted}

\item{right}{'xts' object to be plotted on the right axis}

\item{data}{data object containing \code{...} and \code{right}}

\item{time}{time period to restrict plot. Uses 'xts' convention, e.g.,
'1970-01/1972-06-03'.}

\item{xlim}{Date vector of limits on the x-axis}

\item{ylim}{numeric vector of limits on the left y-axis}

\item{yRlim}{numeric vector of limits on the right y-axis}

\item{ticks}{numeric vector of the desired number of ticks on (x,y) axis}

\item{theme}{graphical theme of the plot. Only 'std' is supported.}

\item{col}{character vector of color names to use for lines}

\item{colBG}{background color}

\item{colGrid}{color of grids}

\item{colAxes}{color of axes}

\item{colPlot}{color of plot background}

\item{colBar}{color of bars (business cycle)}

\item{mar}{inner margins outside of plot. c(bottom,left,top,right)}

\item{bars}{object containing Peaks and Troughs to plot. 'cycle' (default)
plots business cycle bars}

\item{las}{tick label orientation on axis}

\item{cex}{size of objects in plot}

\item{cex.lab}{size of labels}

\item{cex.axis}{size of axis}

\item{lty}{line type. examples: 'solid', 'dotted', 'dashed'}

\item{lwd}{line width}

\item{xlab}{character label of x-axis}

\item{ylab}{character label of left y-axis}

\item{yRlab}{character label of right y-axis}

\item{title}{character title of plot}

\item{ltyAxis}{line type of axis}

\item{lwdAxis}{line width of axis}

\item{lwdTicks}{line width of ticks}

\item{leg}{vector of character names to use in the legend. 'auto' uses
object attributes}
}
\value{
plot of time-series
}
\description{
\code{tPlot} plots economic time series data, taking into consideration
object attributes such as 'units' and 'title'. tPlot works best with data
collecting functions, such as \code{fred} in the the \code{recon} package.
}
\section{Graphical Parameters}{

If \code{NULL}, then value is chosen either by \code{theme} or with object
attributes.
}
\examples{
GDPgap <- fred('gap')
cycU <- fred('cyclical')
GDP <- fred('GDPC1')
tPlot(GDPgap, cycU)
tPlot(GDPgap, right = GDP)
}

