% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{trans}
\alias{trans}
\title{Transform Data}
\usage{
trans(x, type = NULL, opt = NULL)
}
\arguments{
\item{x}{'xts' object}

\item{type}{type of transformation to perform.
\itemize{
  \item \code{log} log transformation
  \item \code{l} lag of time series
  \item \code{g} growth rate of time series
  \item \code{gy} yearly growth rate
  \item \code{d} difference of time series
  \item \code{dy} yearly difference
  \item \code{dv} difference between \code{x} and \code{opt}
  \item \code{pdv} percent diff. between \code{x} and \code{opt}
  \item \code{ldv} log diff. \code{x} and \code{opt}
  \item \code{i} \code{x} is indexed to \code{opt}=100
  \item \code{r} transformed to Real by \code{opt}, defaults to CPI
  \item \code{s} share of \code{opt}, defaults to nominal GDP
}}

\item{opt}{additional options passed to the transformation}
}
\value{
transformed 'xts' object
}
\description{
\code{trans} transforms an object (an 'xts' for all options) by a
specific formula.
}
\examples{
GDP <- fred('GDP')  # Nominal GDP
rGDP <- trans('GDP', 'r', opt = fred('GDPDEF'))
GDPgap <- trans(rGDP, 'pdv', opt = fred('GDPPOT'))

Treas <- fred('GS10')         # 10 year Treasury Yield
rTreas <- trans(Treas, 'r')   # Real interest rate
trans(rTreas, 'i', '2011-01') # indexed to Jan 2011 = 100

}

