% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_functions.R
\name{fred}
\alias{fred}
\title{Download Federal Reserve Data}
\usage{
fred(..., trans = NULL, form = NULL, freq = NULL, date = NULL,
  opt = NULL, name = NULL, units = NULL)
}
\arguments{
\item{...}{vector of FRED codes as strings. If string begins with '?',
\code{fredSearch} is called to search for the series. The following
codes are also accepted: 'gap', 'cyclical', 'inflation', 'Y', 'I', 'G',
and 'NX'.}

\item{trans}{vector of characters used to transform the corresponding
series. Calls the \code{trans} function. Accepts 'log', 'l', 'g', 'gy',
'd', 'dy', 'dv', 'pdv', 'ldv', 'i', 'r', and 's'.}

\item{form}{character string. If specified, formula used to combine the FRED series.
Similar to the 'formula' field on the FRED website. Each series in ...
is assigned a letter (a)-(z) sequentially to be used in the formula.}

\item{freq}{time frequency to convert with which to convert the series.
Accepts 'y', 'q', 'm', 'w', or 'd'.}

\item{date}{starting and ending period for the series. Uses 'xts'
conventions, e.g. \emph{'1970-01/1972-10'}}

\item{opt}{additional options passed to \code{trans}}

\item{name}{vector specifying the resulting column names. If \code{NULL},
the FRED code is used.}

\item{units}{units of the resulting object. Appended to the attributes.
If \code{NULL}, the units from the FRED website are used.}
}
\value{
'xts' object
}
\description{
\code{fred} downloads data from the Federal Reserve Economic Database
(FRED) and formats it to an 'xts' object.
}
\examples{
Y <- fred('GDPC1', trans = 'gy') # yearly growth rate of Real GDP
gap <- fred('GDPC1', 'GDPPOT', form = '(a-b)/b*100', units = 'Percent')
urate <- fred('?unemploy rate')  # evokes search

}

