% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pwt9}
\alias{pwt9}
\title{Penn World Tables 9.0}
\format{A tibble with 11830 rows and 47 variables:
\describe{
  \item{countrycode}{3-letter ISO country code}
  \item{country}{Country name}
  \item{currency_unit}{Currency unit}
  \item{year}{Year}
  \item{rgdpe}{Expenditure-side real GDP at chained PPPs (in mil. 2011US$)}
  \item{rgdpo}{Output-side real GDP at chained PPPs (in mil. 2011US$)}
  \item{pop}{Population (in millions)}
  \item{emp}{Number of persons engaged (in millions)}
  \item{avh}{Average annual hours worked by persons engaged}
  \item{hc}{Human capital index}
  \item{ccon}{Real consumption of households and government, at current PPPs (in mil. 2011US$)}
  \item{cda}{Real domestic absorption (consumption + investment), at current PPPs (in mil. 2011US$)}
  \item{cgdpe}{Expenditure-side real GDP at current PPPs (in mil. 2011US$)}
  \item{cgdpo}{Output-side real GDP at current PPPs (in mil. 2011US$)}
  \item{ck}{Capital stock at current PPPs (in mil. 2011US$)}
  \item{ctfp}{TFP level at current PPPs (USA=1)}
  \item{cwtfp}{Welfare-relevant TFP levels at current PPPs (USA=1)}
  \item{rgdpna}{Real GDP at constant 2011 national prices (in mil. 2011US$)}
  \item{rconna}{Real consumption at constant 2011 national prices (in mil. 2011US$)}
  \item{rdana}{Real domestic absorption at constant 2011 national prices (in mil. 2011US$)}
  \item{rkna}{Capital stock at constant 2011 national prices (in mil. 2011US$)}
  \item{rtfpna}{TFP at constant national prices (2011=1)}
  \item{rwtfpna}{Welfare-relevant TFP at constant national prices (2011=1)}
  \item{labsh}{Share of labour compensation in GDP at current national prices}
  \item{delta}{Average depreciation rate of the capital stock}
  \item{xr}{Exchange rate, national currency/USD}
  \item{pl_con}{Price level of CCON (PPP/XR), price level of USA GDPo in 2011=1}
  \item{pl_da}{Price level of CDA (PPP/XR), price level of USA GDPo in 2011=1}
  \item{pl_gdpo}{Price level of CGDPo (PPP/XR), price level of USA GDPo in 2011=1}
  \item{i_cig}{Is relative price data for consumption, investment, and government extrapolated (0), benchmark (1), or interpolated (2)?}
  \item{i_xm}{Is relative price data for exports and imports extrapolated (0), benchmark (1), or interpolated (2)?}
  \item{i_xr}{Is the exchange rate market-based (0) or estimated (1)?}
  \item{i_outlier}{Is the observation on pl_gdpe or pl_gdpo an outlier? (1)}
  \item{cor_exp}{Correlation between expenditure shares of the country and the US}
  \item{statcap}{Statistical capacity indicator}
  \item{csh_c}{Share of household consumption at current PPPs}
  \item{csh_i}{Share of gross capital formation at current PPPs}
  \item{csh_g}{Share of government consumption at current PPPs}
  \item{csh_x}{Share of merchandise exports at current PPPs}
  \item{csh_m}{Share of merchandise imports at current PPPs}
  \item{csh_r}{Share of residual trade and GDP statistical discrepancy at current PPPs}
  \item{pl_c}{Price level of household consumption, price level of USA in 2011=1}
  \item{pl_i}{Price level of capital formation, price level of USA in 2011=1}
  \item{pl_g}{Price level of government consumption, price level of USA in 2011=1}
  \item{pl_x}{Price level of exports, price level of USA in 2011=1}
  \item{pl_m}{Price level of imports, price level of USA in 2011=1}
  \item{pl_k}{Price level of the capital stock, price level of USA in 2011=1}
}}
\source{
\url{http://www.rug.nl/ggdc/productivity/pwt/}
}
\usage{
pwt9
}
\description{
A dataset containing data from the Penn World Tables version 9.0. It
includes data from 178 countries. The names of each country are listed
in the 'country' data set.
}
\keyword{datasets}
