% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg.R
\name{Reg}
\alias{Reg}
\title{Regression}
\usage{
Reg(form, data = NULL, se = "auto", type = "ols",
  na.action = na.exclude, table = TRUE)
}
\arguments{
\item{form}{regression formula, e.g. \emph{y~x1+x2}}

\item{data}{object that contains data in formula}

\item{se}{type of standard error to use. For OLS regressions, defaults to
'HC1' (robust option in STATA).}

\item{type}{regression type to use. Only 'ols' supported for now}

\item{na.action}{how are NAs handled. Defaults to na.exclude}

\item{table}{logical. Should results be printed}
}
\value{
'RegM' object containing regression information, including
  coefficients, se, tests, and others
}
\description{
\code{Reg} performs a variety of regressions within a single function (only
OLS is supported at the moment). Standard test are automatically performed.
Multiple types of standard error calculations are available.
}
\examples{
GDPgap <- fred('gap')
cycU <- fred('cyclical')
Reg(GDPgap ~ cycU) # Okun's Law
}

